Address reviewer feedback for README and add check for /opt/amazon before removing it

+++ dockerfile_scripts/cray-libs.sh
@@ -6,7 +6,9 @@ set -x
 # libfabric and the AWS plugin. We need to remove it to prevent issues
 # that could occur if that version is loaded instead of the libfabric/ofi
 # libraries we build here.
-rm -rf /opt/amazon
+if [ -d "/opt/amazon" ] ; then
+    rm -rf /opt/amazon
+fi
 
 # Install Cray libcxi. This requires grabbing the cassini/cxi headers
 # and installing them into ${HPC_DIR} so we can compile libcxi.
